Fix #74 Refacto services + namespaces
